Move logic of utxos table to separate function, allow aborting it
