+++ cmd/broadcaster-cli/app/keyset/utxos/utxos.go
@@ -2,18 +2,13 @@ package utxos
 
 import (
 	"context"
-
 	"fmt"
 	"log"
-
-	"sort"
-	"strconv"
-
-	"time"
+	"os"
+	"os/signal"
 
 	"github.com/bitcoin-sv/arc/cmd/broadcaster-cli/helper"
 	"github.com/bitcoin-sv/arc/internal/woc_client"
-	"github.com/jedib0t/go-pretty/v6/table"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -40,117 +35,18 @@ var Cmd = &cobra.Command{
 		if err != nil {
 			return err
 		}
-		t := table.NewWriter()
 
-		type row struct {
-			satoshis string
-			outputs  string
-		}
-		columns := make([][]row, len(keySets))
-		maxRowNr := 0
+		signalChan := make(chan os.Signal, 1)
+		signal.Notify(signalChan, os.Interrupt) // Listen for Ctrl+C
 
-		keyTotalOutputs := make([]int, len(keySets))
-		keyHeaderRow := make([]interface{}, 0)
-		headerRow := make([]interface{}, 0)
-		for i, keyset := range keySets {
+		ctx, cancel := context.WithCancel(context.Background())
 
-			headerRow = append(headerRow, "Sat", "Outputs")
-			keyHeaderRow = append(keyHeaderRow, "key-"+strconv.Itoa(i), "")
+		go func() {
+			<-signalChan
+			cancel()
+		}()
 
-			utxos, err := wocClient.GetUTXOsWithRetries(context.Background(), keyset.Script, keyset.Address(!isTestnet), 1*time.Second, 5)
-			if err != nil {
-				return fmt.Errorf("failed to get utxos from WoC: %v", err)
-			}
-
-			outputsMap := map[uint64]int{}
-			satoshiSlice := []uint64{}
-			var found bool
-			for _, utxo := range utxos {
-				_, found = outputsMap[utxo.Satoshis]
-				if found {
-					outputsMap[utxo.Satoshis]++
-					continue
-				}
-
-				outputsMap[utxo.Satoshis] = 1
-
-				satoshiSlice = append(satoshiSlice, utxo.Satoshis)
-			}
-
-			sort.Slice(satoshiSlice, func(i, j int) bool {
-				return satoshiSlice[j] < satoshiSlice[i]
-			})
-
-			totalOutputs := 0
-
-			for _, satoshi := range satoshiSlice {
-
-				columns[i] = append(columns[i], row{
-					satoshis: strconv.FormatUint(satoshi, 10),
-					outputs:  strconv.Itoa(outputsMap[satoshi]),
-				})
-
-				totalOutputs += outputsMap[satoshi]
-			}
-
-			keyTotalOutputs[i] = totalOutputs
-
-			if len(columns[i]) > maxRowNr {
-				maxRowNr = len(columns[i])
-			}
-		}
-
-		t.AppendHeader(keyHeaderRow)
-		t.AppendHeader(headerRow)
-
-		rows := make([][]string, maxRowNr)
-
-		for i := 0; i < maxRowNr; i++ {
-			for j := range columns {
-
-				if len(columns[j]) < i+1 {
-					rows[i] = append(rows[i], "")
-					rows[i] = append(rows[i], "")
-					continue
-				}
-
-				rows[i] = append(rows[i], columns[j][i].satoshis)
-				rows[i] = append(rows[i], columns[j][i].outputs)
-			}
-		}
-
-		for i, row := range rows {
-			tableRow := table.Row{}
-
-			if maxRows != 0 && i == maxRows+1 {
-				for range row {
-					tableRow = append(tableRow, "...")
-				}
-
-				t.AppendRow(tableRow)
-
-				continue
-			}
-
-			if maxRows != 0 && i > maxRows {
-				continue
-			}
-
-			for _, rowVal := range row {
-				tableRow = append(tableRow, rowVal)
-			}
-
-			t.AppendRow(tableRow)
-		}
-
-		totalRow := table.Row{}
-		for _, total := range keyTotalOutputs {
-			totalRow = append(totalRow, "Total", total)
-		}
-		t.AppendRow(totalRow)
-
-		// Todo: Add row with total Satoshis
-
+		t := getUtxosTable(ctx, logger, keySets, isTestnet, wocClient, maxRows)
 		fmt.Println(t.Render())
 
 		return nil
